Use Realname switch for edit Shift User Dropdown
Issue86 in engelsystem
There was already one function, just have to call it.
Its not best practise to use <a> in select options,
but it works and there is no better function to use.

This fixes only one certain dropdown. Its still possible it could
be somewhere else as well. Not sure if this Issue means to find
and fix them all.